Перенес методы из MdoReferences в существующий класс MdoRefBuilder
и добавил использование стринг интернера
